IdleDetector: Use AbortSignal::Algorithm
It is better to use AbortSignal::Algorithm for adding "abort algorithms"
to the AbortSignal as introduced in https://chromium-review.googlesource.com/c/chromium/src/+/3302090.

This is because then we do not need to worry about the weakness of the
AbortSignal in IdleDetector::Abort.

Bug: 1309731
Change-Id: Idf238df4dd13219fffef7a9228a7297c9e3374bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3558659
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987292}
NOKEYCHECK=True
GitOrigin-RevId: 7dbc145242b3919892dcf6cb39287f408c9b3bee