[blinkpy] Raise xvfb maxclients in more cases
This is a reland of [1].  In  my original CL, I thought we had no more
Trusty bots left, but it turns out linux-trusty-rel is still around.

This CL adds a runtime check for -maxclients support for xvfb.  This
matches the behavior in testing/xvfb.py [2].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3550636
[2] https://source.chromium.org/chromium/chromium/src/+/main:testing/xvfb.py;drc=9739e177315123f291cb2b455a7d1cca6415fdee;l=173

R=smcgruer

Bug: 1223431
Change-Id: I920e2fa5f4bf47b6efe192c1ebbb98aef9dbb7bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3552390
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987487}
NOKEYCHECK=True
GitOrigin-RevId: b08a473ac68f913f4cc0d5f69263c98cc457c41e