CanvasFormattedText Styling Prototype
This change resurrects @sushraja's changes to add styling abilities to
CanvasFormattedText.  Note that this implements an older version of the
API, which is changing rapidly.

Original description:

Add support for styling CanvasFormattedText

This change adds a styleMap on the CanvasFormattedText and
CanvasFormattedTextRun. A subset of CSS properties are supported on the
CanvasFormattedText and a smaller subset on the CanvasFormattedTextRun.

A height parameter is added to fillFormattedText so that vertical
writing modes can be supported. The height/width passed to
fillFormattedText are used as the available space, while the
CSS height/width can be set on the CanvasFormattedText to control
height/width explicitly.

Change-Id: I2af3e2bc667f4d6e57da390fdf0648663a75461e
Bug: 1155764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3491151
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Ian Prest <iapres@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#987620}
NOKEYCHECK=True
GitOrigin-RevId: dac79e04ef1c10f64cd7a0117cc57b61b5be1f30