Remove/update old beforematch events
I forgot to remove the old beforematch events for scroll to text
fragment and element fragment navigation when I migrated everything from
content-visibility:hidden-matchable to hidden=until-found. The old
events are fired regardless of the hidden-matchable or
hidden=until-found, but the specced version of the feature says that the
beforematch event is only fired on hidden=until-found elements.

I also tried to remove the async step for beforematch from scroll to
text fragment, but TextFragmentAnchor::Invoke is forbid from running
script, so we at least still need the animation frame callback in order
to run the revealing stuff.

This patch also updates and consolidates the beforematch tests for
element fragment navigation and moves them to WPT. These are the only
beforematch tests which can be included in WPT because scroll to text
fragment is not in the HTML spec yet and find-in-page can't be tested in
WPT.

Bug: 1055002
Change-Id: I55043f37a167d5fa222564a92eff3bb80e4c8d85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3558274
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987650}
NOKEYCHECK=True
GitOrigin-RevId: f6914d1e4a1e355d437715675b652032f45fd4c9