Image::ResizeAndOrientImage: Do not convert to sRGB by default.
This function converts the input image to 8-bit sRGB. This conversion
is undesirable for most of its callers.

Add an optional parameter through which a target color space may
be specified. If not specified, use the input PaintImage's color
space. Specify this target color space as sRGB for DrawImage, because
DrawImage discards color space information and assumes sRGB.

Draw out in comments that the conversion also converts to
kN32_SkColorType. This is also likely undesirable, but is not
severely breaking things just yet.

This unblocks respecting WebGL's unpackColorSpace, which was clamping
non-sRGB inputs to sRGB because of this function's use inside
WebGLRenderingContextBase::TexImageHelperImageBitmap.

Bug: 1310811
Change-Id: Ie73235b98cdf88e32d1ae55df2a11a9e400302dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3562270
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987664}
NOKEYCHECK=True
GitOrigin-RevId: 42c4e8186db748b0b037ad64a55e124cb7fcc13d