Pre-paint: Don't miss AREA elements in culled inlines.
We'd miss them when block-fragmented. We used to assume that the parent
of the AREA would occur in the last block fragment, but that's not a
safe assumption to make.

Instead, enter it when visiting the parent for the first time. We only
have a flag for the first time visiting, not for the last time, but it
doesn't really matter, as long as we visit it once, and only once.

Bug: 1308078
Change-Id: If549ad14b81e683af47b08c13fe42747c853e5a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3562999
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987805}
NOKEYCHECK=True
GitOrigin-RevId: 3ce7ef782fe8191121e750949a20da4e9da25dd9