WebGL: Convert ImageBitmap to unpack color space
When uploading an ImageBitmap using texImage2D, ensure that the contents
be converted to the unpack color space before uploading.

Do this by calling StaticBitmapImage::ConvertToColorSpace inside
WebGLRenderingContextBase::TexImageHelperImageBitmap.

Change several uses of ImageBitmap to operate on the color converted
StaticBitmapImage (because the ImageBitmap references the not-converted
StaticBitmapImage). As part of this, move the helper function
CopyImageData to be a member function of StaticBitmapImage.

This should have almost no effect, because the unpack color space is
sRGB, almost all ImageBitmaps are (inappropriately) pre-converted to
sRGB.

Bug: 1310811
Change-Id: I5632c79f76d509212aab4848930b616e37d91532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560989
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987932}
NOKEYCHECK=True
GitOrigin-RevId: 18853783cb83d8b12f67b98e2cd8e3656f1d3250