Deny notification/push permission for documents in non-standard StoragePartitions
This CL blocks notification permission for documents that are using a
different StoragePartition than they would if they were loaded in a top
level frame. This can happen for documents iframed in a page that uses
a non-default StoragePartition. Permission delegation typically handles
this case, but there are a few non-delegated permissions that would
break in this case because permissions aren't scoped by StoragePartition.

This CL defines a new PermissionManager::GetPermissionStatusForWorker
method that takes a RenderProcessHost and origin, and can determine if
the process is using its "home" StoragePartition. All worker-initiated
notification permission checks are migrated to this new method, and
eventually all worker-initiated permission checks should use it.

See crbug.com/1266685 for more details.

Bug: 1266685
Change-Id: I401d5b00ba0a19302842a917feff3f18d71def0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3257305
Reviewed-by: Illia Klimov <elklm@chromium.org>
Reviewed-by: Michael Bai <michaelbai@chromium.org>
Reviewed-by: Junbo Ke <juke@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987989}
NOKEYCHECK=True
GitOrigin-RevId: 8d5602ac5d60468e1a28db48bd6cd761bafebe13