Ensure fallback setup when manual slots are inserted into shadow DOM
When a new slot is added to a shadow DOM with manual slot assignment,
while the slot is guaranteed to not have any assigned node, it may still
have fallback content. This patch ensures that the fallback content is
correctly set up, so that it can be correctly rendered, and can be
correctly cleared when the slot gets an assignment.

Test cases:
- imperative-slot-initial-fallback.html: Tests that the fallback content
  of an imperative slot is correctly rendered if it has no assignment
- imperative-slot-fallback-clear.html: Tests that after rendering the
  fallback content and then getting an assignment, an imperative slot
  can correctly clear the fallback content

Fixed: 1292292
Change-Id: I74403ab7c6003a9a0203cfdce1c2896f9529499e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3562222
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988106}
NOKEYCHECK=True
GitOrigin-RevId: 53a781076cce05d820f0bdc2fdde3064cae1ba79