[FLEDGE] Add update URL to generateBid()'s interestGroup parameter.
As required by spec. To do this, the update URL is added to the
BidderWorkletNonSharedParams. Unlike most of the other parameters,
it's only used to directly set the corresponding interest group field in
the generateBid() call.

Also rename `update_url` in InterestGroup-related structs to
`daily_update_url` to match the Javascript name of the field.

Bug: 1311052
Change-Id: I67045a72f50a186e51f6a5225c28a5a8d629ac8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3554853
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988149}
NOKEYCHECK=True
GitOrigin-RevId: 4c587c463b4dc37b6ad040c894183ed0c4a2b794