Stop triggering bfcache eviction from loaders with wrong freeze_mode_
BFCache eviction should only be triggered when the freeze mode is
WebLoaderFreezeMode::kBufferIncoming. If it's kNone, then page is
not in BFCache, and if it's kStrict, eviction is disallowed as the page
is being restored and it's too late to evict.

WebResourceRequestSender::SendSync never runs when the loader is frozen,
so it never runs when the page is in BFCache. This CL removes the
eviction call from that function, ensures
MojoURLLoaderClient::EvictFromBackForwardCacheDueToTimeout() is only
called when the freeze mode is kBufferIncoming, and adds DCHECKs in
various loader eviction functions to assert the freeze mode.

Bug: 1305041
Change-Id: Iac5a0391500be2be53d54dd4e1990558ebb28042
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3531689
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988436}
NOKEYCHECK=True
GitOrigin-RevId: 3d0d19c89151053cc7e148321634a6c70d0922e3