Add document-wide cache for StyleFetchedImage
StyleFetchedImage is cached on the CSSImageValue object and re-used each
time a declaration applies to an element's ComputedStyle. The problem
arises when the same url is used in different declarations, or even
worse, when the CSSImageValue is regenerated every time when a custom
property url is substituted into another property. In those cases we
create a new StyleFetchedImage and re-fetch the resource.

Use the same StyleFetchedImage for the cached image url for all
CSSImageValue objects for the same Document. The cache is a map from url
into a weak StyleFetchedImage such that it will be garbage collected
once there are no more CSSImageValues referencing the StyleFetchedImage.

Bug: 961600
Change-Id: I5978c9b6befc9b0aa2f01e7e764866e2d80a8011
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3565368
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988488}
NOKEYCHECK=True
GitOrigin-RevId: 288ae84c98bb0b40cb5480a2958f074f7cb0a3ac