Check for null window in PaymentInstruments
Add a check for a null window in PaymentInstruments::set, which is the
case when called by a service worker. In this case, there is no window
to check for a user gesture.

This is a quick fix for the null dereference, added a TODO to consider
next steps here.

Bug: 1311953
Change-Id: Ib3f8262ddeee3d5fdb0ddefbeee26149921ecbca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3564418
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988533}
NOKEYCHECK=True
GitOrigin-RevId: 21e746736039f08d0a9727a422f9d72e85452324