Check feature policy in PaymentAppServiceWorkerRegistration
Add a renderer-side check for PaymentAppServiceWorkerRegistration for
the payment feature policy, and throw an exception if it's not allowed.
This is also checked on the browser side, but the browser kills the
renderer if it's violated, so the renderer should check first and raise
an exception.

The AllowedToUsePaymentFeatures logic to check the feature policy comes
from payment_instruments.cc.

Bug: 1311949
Change-Id: I604ab5ef22cbe894a1ff7f578d4a28d622e2418d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3563759
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988544}
NOKEYCHECK=True
GitOrigin-RevId: c3e869153c533d0a470da37cafa9c80ba68ce1d3