Disable relayout on intrinsic size change when fragmented.
The particular problem that I found with this was the call to
ClearLayoutResults(), which ended up removing not only the layout result
we had just produced, but any results from previous fragmentainers as
well, causing misery further down the road when trying to write back the
new fragment to the LayoutBox at an index out of range. Since it doesn't
seem to make a lot of sense to handle intrinsic inline-size changes
after having produced fragments in earlier fragmentainers, just disable
this if we're resuming layout.

See also crbug.com/1312948

Bug: 1308960
Change-Id: I7ab97507420473bc6c3522d75128ae3292cf3e26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3568457
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988664}
NOKEYCHECK=True
GitOrigin-RevId: 132ddbc5dfa6864c3b8224ed8f6d3975e40e4454