Fix _LintWPT for Windows command limits
_LintWPT executes a script on all applicable files. When running
"git cl presubmit --all" that ends up being ~65,000 files and the
the command line that is generated is  over 7,000,000 characters.
Windows does not like that.

This change breaks up the check into multiple invocations of the script
in order to avoid these limits. This is important because running all
presubmits is the only way to systematically find bugs, and avoid being
surprised by them when submitting a change.

See crrev.com/c/3554633 for another example of fixing this issue.

Running this command:

    git cl presubmit --force --files=*-expected.txt

is sufficient to trigger this code and therefore test it, and this was
run on this CL, with crrev.com/c/3564682 patched in to avoid other
errors.

Bug: 1309977
Change-Id: I6db0c95ee92e689ddddd0b8515f459155b766c77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3564185
Reviewed-by: Jesse McKenna <jessemckenna@google.com>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988674}
NOKEYCHECK=True
GitOrigin-RevId: cd3cfa81dca0cc46aaee8626deddd411a9596859