media::AudioProcessor: Don't use an EchoControlFactory when building WebRTC APM
The code currently supplies an echo controller factory to WebRTC, which overrides built-in configuration for stereo playout added in this CL:
https://webrtc-review.googlesource.com/c/src/+/161086

With this CL, Chrome benefits from the stereo configuration.

As a safety measure, the stereo configuration has a killswitch that can be used if this CL unexpectedly causes issues for users:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/webrtc/modules/audio_processing/audio_processing_impl.cc;l=62-68;drc=7fb345a0da63049b102e1c0bcdc8d7831110e324;bpv=1;bpt=1

Some code (echo control instantiation within the WebRTC APM) is not guarded by any killswitch.

Tested: Modified Chrome to run ChromeWideEchoCancellation and print AEC3 config. Verified in test call that AEC3 switched to stereo support with this CL but did not without this CL.
Bug: chromium:1312870,webrtc:11151,chromium:1029717
Change-Id: I699d6f3f1d268d2c869f107f610f1a8a8115c6c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3563402
Reviewed-by: Per Åhgren <peah@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Sam Zackrisson <saza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988948}
NOKEYCHECK=True
GitOrigin-RevId: d904c26cf6a2b3f316dced8198ea4c56dbb5f9d5