Fenced frames: Browser side reporting metadata
This CL implements the browser process side changes for the events
reporting design. The browser side changes require reporting related
metadata to be stored for a given urn:uuid and passed to the renderer
on navigation commit. The subsequent CLs will handle the renderer side
design which when reportEvent() is invoked by the fenced frame will
send a beacon to the url part of this metadata.
This CL tests the metadata is stored successfully in the mapping and the
subsequent CLs that use in the renderer will test the end-to-end flow.

The reporting explainer can be found here:
https://github.com/WICG/turtledove/blob/main/Fenced_Frames_Ads_Reporting.md

and the implementation design is here:
https://docs.google.com/document/d/1cpz0ulw363AKRcQIbWEItPA4iGb05WZZqf2c57AZ-eo/edit?usp=sharing

Change-Id: I257c293697df06bfef720480c16377344cddafe2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3535041
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989108}
NOKEYCHECK=True
GitOrigin-RevId: c7b97ca9e9cfde98785d8a30fe3a8aefb5ab8407