[LayoutNG] Fixedpos in abspos in inline container
A fixedpos nested in a fragmented OOF would crash when its CB was
an inline because we would set the
fixedpos_containing_block of the OOF ancestor to the inline.
However, the containing block should not be set to an inline, but
rather to the closest parent block flow of the inline.

To fix this, add a new member to the OOF node to keep track of
a fixedpos_inline_container and set the fixedpos_containing_block
once the fixedpos_inline_container's parent block flow is reached.
This allows us to make use of the same logic for handling
inline CBs as we would for normal un-nested OOFs.

The same issue also occurs for nested multicols, so relevant
crashing tests have been added and a similar
fixedpos_inline_container member was added to
NGMulticolWithPendingOOFs.

Ref tests were not added as a part of this CL, although
I suspect the relative offset logic might be off. TODO added to
investigate this separately.

Bug: 1312790
Change-Id: I236006a74e8917745cbaacbe51fc71b7e393f11b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3573056
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#989483}
NOKEYCHECK=True
GitOrigin-RevId: b955e1e7bae070f86af0f516037bc6575e375d4d