Remove IsAnonymousNGMulticolInlineWrapper LayoutObject bit.
Instead, calculate whether an anonymous block is a direct child of the
multicol container (ignoring the flow thread) on the fly.

We could probably have fixed this by having LayoutBlockFlow::
MakeChildrenNonInline() update the bit as well, but this just seems too
tricky (and MakeChildrenNonInline() might not even be the only place
that was missing this).

Bug: 1308071
Change-Id: I16090a9252590b346dfaf27d34a798347710b704
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3571876
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989503}
NOKEYCHECK=True
GitOrigin-RevId: 1cb406004cf22a7813be151cdb28765022ac797d