Check ContextIsValid() for workers in ToScriptState()
To make the code of ToV8Context() and ToScriptState() aligned.

This CL also adds a null check on ScriptState to avoid
potential crashes.

This CL doesn't change the behavior, because the ScriptState
is always valid for workers at the existing call sites.

Bug: 1313328
Change-Id: Idc2ed27330d113372c43b4461f96d076bafb4fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469542
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Hiroshige Hayashizaki <hiroshige@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989553}
NOKEYCHECK=True
GitOrigin-RevId: 0aa199ae4077345025a7f8a0cca4dc0130dca528