Use layout coordinates for LayoutClipRect of ClipPathClip and MaskClip
This is to ensure correct inclusive intersection for empty objects
with empty clip-path/mask. Previously, we failed to detect viewport
intersection and start loading a lazy loading <img> which was empty
with an empty clip-path before the image was loaded.

A change (crrev.com/936552) from blink::EnclosingIntRect(FloatRect) to
gfx::ToEnclosingRect(gfx::RectF) exposed the issue. The difference
between the two functions was that for an empty input rect with
non-integral origin, the former returned an non-empty IntRect, while
the latter returns an empty gfx::Rect. An alternative to this CL is to
add a version of gfx::ToEnclosingRect() behaving the same as
blink::EnclosingIntRect(), but I think this CL is better by making
LayoutClipRect for ClipPathClip and MaskClip correct.

Bug: 1308299, 1248598
Change-Id: I242c1d38277cdd91aa596fda15156d991f2626d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3568746
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989750}
NOKEYCHECK=True
GitOrigin-RevId: cc453fd9995beb22738596955bb9fde75054db18