Fix computed style for currentcolor on highlight pseudos
Computed style for "currentcolor" on highlight pseudos
has to return the color of the originating element.

This patch passes the originating element style
on the StyleRequest, and uses it in Color::ApplyValue().

Apart from that, we set ColorIsCurrentColor flag properly
in Color::ApplyValue(), as right now we were losing it
if the parent style didn't have "currentcolor" too.

Bug: 1295264
Change-Id: I31e7ffb6cfeafaa54da4f3a9e21eff2babbdaa3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3563099
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Manuel Rego <rego@igalia.com>
Cr-Commit-Position: refs/heads/main@{#989834}
NOKEYCHECK=True
GitOrigin-RevId: 368ce3dae0e9944607968f72d180c9c52e2ca315