Create MediaStreamAudioDestinationNode's Source and PlatformSource together
Part of a broader effort to align the lifecycles of MediaStreamSource
and the platform source set on it, in this case a
WebAudioMediaStreamSource.
We've been working through removing calls to MSS::SetPlatformSource,
instead providing it in the constructor.
For MediaStreamAudioDestinationNode (MSADN) this requires moving the
creation of the WebAudioMSS to the constructor of MSADN, rather than
indirectly via the call to DidCreateMediaStreamAndTracks() and eventually
MediaStreamUtils::CreateNativeAudioMediaStreamTrack().

#removeplatformsetters

Bug: 1302689
Change-Id: I6eb2ac33c1eb596f87346ecd8bed3242d6a6c080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3571833
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Auto-Submit: Tony Herre <toprice@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989865}
NOKEYCHECK=True
GitOrigin-RevId: 29ceacc4d7044ec525bdd197534fdfea9093f060