Speculation Rules: Remove rule sets when <script> is removed.
Changes made in the same task are guaranteed to be reported atomically.
Changes made after a previous report will cause a second report to
the browser process.

Tests for both of these cases are included. The intermediate case is
ultimately a scheduling issue, but will land in one of these two cases.

Bug: 1314003
Change-Id: Ic3b125d0046600020018e9b4bc8b56d59e1dd781
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3573461
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989949}
NOKEYCHECK=True
GitOrigin-RevId: f427fc4dd14663e990e3d73ef28c18dee653aef5