Reland "Change kCanvas2dStaysGPUOnReadback check to FeatureList::IsEnabled"
This is a reland of commit ee8802862f66915f581158bec734026b68482d80

It was revered here https://chromium-review.googlesource.com/c/chromium/src/+/3550238

For users with the experiment (which is currently 50% of dev and canary) there will be performance regressions for motionmark scores. No one on release is yet seeing this. See the crbug.com/1288118 for more details, please do not revert solely for this reason.

In short: with the new getContext(`2d`, {willReadFrequently: true}) there is no regression. Unfortunately motionmark will not update their context creation until it is more widely used on the web. This experiment is part of an effort to get the feature more widely used and convince motionmark to update.

Original change's description:
> Change kCanvas2dStaysGPUOnReadback check to FeatureList::IsEnabled
>
> Apparently if you don't do it this way it doesn't appear in the finch dashboard.
>
> Gonna get rid of the RuntimeEnabledFeatures version because it seems like a bad idea to have two versions of the same flag. Since `experimental-web-platform-features` seems to only work on RuntimeEnabledFeatures, I'm just going to enable this directly for the two tests that need it.
>
> Bug: 1288118
> Change-Id: Ib91faeea8c93e00672d9e50252c416c5b6466ffc
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3526073
> Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
> Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
> Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
> Reviewed-by: Charlie Reis <creis@chromium.org>
> Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#983833}

Bug: 1288118
Change-Id: I7be092ff68a8788dcff737cd3578546e42e764eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3570061
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#990151}
NOKEYCHECK=True
GitOrigin-RevId: 70274d596af204ef699574357b80b549126da4c1