[FlexNG] IsBreakableAtStartOfResumedContainer for flex rows
We used to use the first child layout result of a row to determine
IsBreakableAtStartOfResumedContainer() for the row. However, since
items can break before multiple times, a row might not be the
first for the node while the first item in the row is.

Pass is_first_for_node as an input to
IsBreakableAtStartOfResumedContainer() so that we can use the value
specific to the row instead of its item.

The test I added works with and without this CL, but I wasn't able
to think of a case where this actually broke anything.

Bug: 660611
Change-Id: I9cbbd61c383dac0ce149d9f33f8c2b82b983a1d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3575366
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#990455}
NOKEYCHECK=True
GitOrigin-RevId: 5962577ca703742437faeba718fbfb0eed9ab4fe