Promote ImageBitmaps to use HW overlay upon presentation on MacOS
With this change, when an image bitmap's underlying SharedImage is
dispatched to the compositor via ImageLayerBridge, we copy it
to a new shared image if the original one is not an overlay
candidate.

On other platforms (non-MacOS) we do not perform this copy because
it may not be worthwhile.

BUG=1313453

Change-Id: I4707c7b2967b5f31982aa0328869f13625c54e28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3572505
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#990487}
NOKEYCHECK=True
GitOrigin-RevId: e46ff9a03801df2bfff8f094f04d08997f0d84ca