[layout] Fix performance issue with tables.
In https://chromium-review.googlesource.com/c/chromium/src/+/3534720
I moved the calculation of baselines into the a pre-step within
the table layout algorithm.

This still used the kLayout slot - which was incorrect as it bypassed
the tree-structure so in a followup I added a disable side-effects
scope to account for this.

This defeated the caching logic, and as such caused a performance
regression.

This patch effectively reverses the prior patches but now that we've
reworked the logic a little is a little cleaner than what was there
before.

Bug: 1309368
Change-Id: I98deea5ad4b4776c4ab0c5ae246167e9d439116d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3574812
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#990561}
NOKEYCHECK=True
GitOrigin-RevId: f72943e29d0b57758c48b96ecdb5a060e4aaefb2