Replicate Active state to render process for all RenderViews.
The active state was incorrect for child RenderViews of a frame tree
because it was not replicated across the entire page. Ensure that the
active state is indeed replicated.

To accomplish this correctly change RenderWidgetHostViewAura/Mac to
let the WebContentsImpl know about replicating the active state.

RenderWidgetHostViewAndroid does not yet have support for non-active
windows (it would need to be added for Android multi-window support
and likely hook into the onTopActivityResumedChanged) so whenever a
frame receives focus it becomes active. This support is unchanged.

Add is_active to the InputHandler IPC (via compositor thread) because
this can race against the SetActive IPC which is associated with the
legacy IPC channel.

Augment a RenderWidgetHostViewAura test to ensure that a cross site
iframe child has its active state set correctly.

BUG=1313121

Change-Id: I3932147eccd97bbd256d1c0161d8366f781ffdf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3570062
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#990585}
NOKEYCHECK=True
GitOrigin-RevId: 3450d0bba4ea5fe2e44814abd18c6d170b85d1a5