Improve RuntimeEnabledFeaturesBase::SetFeatureEnabledFromString
We call this function numerous time when initializing a renderer process

During the first 30s of renderer process startup we spend (on Android)
at least 0.3% of time within:
content::SetRuntimeFeaturesDefaultsAndUpdateFromArgs

and of that around 0.15% (or around half) within:
RuntimeEnabledFeaturesBase::SetFeatureEnabledFromString

Previously we'd iterate linearly over all the features doing a string
comparison to find the correct feature to set.

This changes the implementation to ensure the feature list is sorted
and perform a binary search instead.

Change-Id: I2c3bbac06e7dfcd11af4f931672b2e4f9d95bdd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3579643
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#990651}
NOKEYCHECK=True
GitOrigin-RevId: 78d4e7c75407b27de4ba2394911fda1b7a8ca5a5