Remove the CascadeExpansion state machine.
The cascade expansion is a very hot part of style application, but does
very little work for each iteration. Combined with the fact that it is
frequently run with very few elements in the array (as many as one),
which makes the construction take undue amounts of resources, the
overhead becomes too significant to ignore.

We replace it with a fairly simple loop, with a lambda callback.
(Nothing in the actual algorithm changes, we just keep state around
in registers instead of in a temporary object with lots of non-inlined
functions.) We need to change the interface and put it into an -inl.h file,
but actually end up with less code than before. We also get a speedup
of about 4% in style, it seems, although cache misses on CSSProperty
is still a significant problem for us.

Change-Id: I2eb7571b10109789461a237a36cfae97bed4c15b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3574574
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#990956}
NOKEYCHECK=True
GitOrigin-RevId: 0a9931d687a999a00d477417143c226079237b5b