Simplify CascadeMap::Add().
Instead of using a lambda that we call from two different places
(which the compiler seemingly refuses to inline, because it is
too big), rewrite into finding the right list first and then
inserting into it in a common block afterwards. This is mainly
a cleanup, but seems to help ~1% in style since we don't have to
go through the function call overhead and Add() is fairly hot.

Change-Id: I72bf01a8dbac7c2cdd7c29eafd90dd714ef40eb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3578697
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#990981}
NOKEYCHECK=True
GitOrigin-RevId: aec4e5fc8ec5c59df3edb290a4f6fba657c832e7