WPT for opener navigations with COOP.
COOP is used to sever relationships with openers, protecting from
side-channel attacks. This can happen both when opening popups and when
navigating, and that essentially relies on the same mechanism.

When a popup navigates, we do all the opener clearing very nicely and
everything ends up in a good state. When the page that opened the popup
navigates instead, we do not have as much coverage and some exploits
were discovered on Chrome (see associated bug, and fix patch:
https://chromium-review.googlesource.com/c/chromium/src/+/3532010)

This patch adds minimal coverage.

Bug: 1305394
Change-Id: I0d1158acf8ba4521eba272c2a9d6170f60b8bd94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3578744
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991150}
NOKEYCHECK=True
GitOrigin-RevId: b465304f25945922b9cf68b7e8bf74c056871ecf