[ChromeVox] Restructure file directory
This change organizes the file directory so that each subdirectory with
JS files (except common) represents a different script context -- such
as the background, the panel, the options page, etc. In a follow-up CL,
all references to scripts that are not in the same directory or common/
will be migrated to use intermediate bridges. This is necessary for the
migration to Manifest v3.

AX-Relnotes: n/a.
Bug: 1288634
Change-Id: I2c5787f0ce4f180b6229da25478e714e222c8b3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3575594
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/main@{#991268}
NOKEYCHECK=True
GitOrigin-RevId: 226be9fb04dbd39fb227ada3cf385017c91bc061