[a11y] Expose Media "more options" button as menu button
The MediaControlOverflowMenuButtonElement was missing the
aria-haspopup=menu attribute. This CL fixes it.

Without this attribute, ATs users won't know that a popup is supposed to
show up when the button is pressed. To fix it, we simply need to add
the "haspopup=menu" attribute on the element.

menu button, helping screen readers announcing that the button is hiding
a collapsed menu of options.

Bug: 1314898
Change-Id: I6b6476c4b0f1bf06685ef31d31322e8ec28efc3a
AX-Relnotes: The media "More options" button is now exposed as a menu
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3508289
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#991372}
NOKEYCHECK=True
GitOrigin-RevId: b387036496309b17e4cfeb130117ddadffd31be3