SVG default behavior in Forced Colors Mode
We had resolved to update the default forced-color-adjust value for
SVGs from none to preserve-parent-color. However, we have been blocked
on shipping this behavior as a result of
https://github.com/w3c/csswg-drafts/issues/6773.

We have received signals that it would be useful to have this work
for SVGs in the meantime while we wait for a CSSWG resolution.

As a temporary fix, allow preserve-parent-color to parse if we are
in the UA stylesheet so that the default behavior of SVGs is updated
to what authors expect without having to ship the preserve-parent-color
property.

Bug: 1164162
Change-Id: I16fea791a376078b5960801eae6190c44499e565
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3580054
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#991577}
NOKEYCHECK=True
GitOrigin-RevId: bd29a99c80f79a969e530b581e18275229c0a53c