Small refactoring in PaintLayer
SetNeedsCompositingInputsUpdate() only calls
MarkAncestorChainForFlagsUpdate() these days.
Removed the argument in that call, because it's not needed
it's the default value).

Apart from that, replaced calls to SetNeedsCompositingInputsUpdate()
on PaintLayer, by MarkAncestorChainForFlagsUpdate().
Also removed a call to SetNeedsCompositingInputsUpdate()
in PaintLayer::AddChild() as it was already calling
MarkAncestorChainForFlagsUpdate().

This is just a refactoring and should have no behavior change.

Change-Id: I9beafc0bdbeb71e46be3c4d0831b620df5782179
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582122
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Manuel Rego <rego@igalia.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991585}
NOKEYCHECK=True
GitOrigin-RevId: 5ecacfeb397b6b0525e376f8d1d21a04119e9cb4