Add feature to establish GPU channel asynchronously
RequestNewLayerTreeFrameSink() calls EstablishGpuChannelSync(), which
can be very slow if the gpu process has not started, or the thread where
it is handled is busy. This can delay navigation commit and possibly
slow down page load. This change adds a feature to experiment with
establishing the gpu channel asynchronously to see if it has any
performance benefits.

Bug: 1278147
Change-Id: I26cebe7aa87affe553f2de3094c7202e9d17032c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3579494
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991605}
NOKEYCHECK=True
GitOrigin-RevId: c734d11b28cb2a389a24e09a138c179d90830cc1